lab_05/pyproject.toml
[tool.poetry]
name = "lab-05"
version = "0.1.0"
description = ""
authors = ["ivaaahn <ivahnencko01@gmail.com>"]
readme = "README.md"
packages = [{include = "lab_05"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
rsa = "^4.9"
pycryptodome = "^3.15.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
